Introduced start() to listener and invocation services. Linked proxy base to listener service.
