+++ hazelcast/proxy/base.py
@@ -34,8 +34,8 @@ class Proxy(object):
         self.logger = logging.getLogger("HazelcastClient.%s(%s)" % (type(self).__name__, name))
         self._to_object = client.serialization_service.to_object
         self._to_data = client.serialization_service.to_data
-        self._start_listening = client.listener.start_listening
-        self._stop_listening = client.listener.stop_listening
+        self._register_listener = client.listener.register_listener
+        self._deregister_listener = client.listener.deregister_listener
 
     def destroy(self):
         """
