+++ start-rc.sh
@@ -6,7 +6,7 @@ else
     USER=""
 fi
 
-HZ_VERSION="3.12.1-SNAPSHOT"
+HZ_VERSION="3.12.1"
 
 HAZELCAST_TEST_VERSION=${HZ_VERSION}
 HAZELCAST_ENTERPRISE_TEST_VERSION=${HZ_VERSION}
